// gge-tracker-backend-api/src/api/enums/http-status.enums.ts
/**
 * Represents standard HTTP status codes used in API responses.
 *
 * @enum {number}
 */
export enum Status {
  OK = 200,
  CREATED = 201,
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  FORBIDDEN = 403,
  NOT_FOUND = 404,
  TOO_MANY_REQUESTS = 429,
  INTERNAL_SERVER_ERROR = 500,
  SERVICE_UNAVAILABLE = 503,
}